services/requirements/requirements.txt
beautifulsoup4~=4.7.1
click==7.1.2
email-validator==1.1.0
flask==1.1.2
flask-bcrypt==0.7.1
flask-login==0.5.0
flask-mail==0.9.1
flask-sqlalchemy==2.4.1
flask-wtf==0.14.3
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
MechanicalSoup==0.12.0
pillow==7.1.2
pycparser==2.20
pymongo==3.8.0
requests==2.22.0
six==1.14.0
sqlalchemy==1.3.16
werkzeug==1.0.1
wtforms==2.3.1
xlrd==1.2.0